+++ client/src/component/dashboard/__tests__/index.js
@@ -18,6 +18,9 @@ describe('Dashboard', () => {
         datafetch: new Map(),
         dashboard: {
           clusters: []
+        },
+        events: {
+          events: []
         }
       });
       const tree = renderer.create(